Activity balance general funcional

+++ app/src/main/java/com/example/vendeton/Entidades/BalanceGeneral.java
@@ -1,11 +1,11 @@
 package com.example.vendeton.Entidades;
 
 public class BalanceGeneral {
-    public BalanceGeneral(String tip_nombre, float monto) {
+    public BalanceGeneral(String tip_nombre, double monto) {
         this.tip_nombre = tip_nombre;
         this.monto = monto;
     }
 
     public String tip_nombre;
-    public float monto;
+    public double monto;
 }
